+++ src/lava/lib/dl/decolle/__init__.py
@@ -1,0 +1,5 @@
+from . import loss
+
+__all__ = [
+    'loss'
+]
