+++ src/lava/lib/dl/decolle/utils/__init__.py
@@ -1,0 +1,5 @@
+from .assistant import DECOLLEAssistant
+
+__all__ = [
+    'DECOLLEAssistant'
+]
